requirements.txt
google-generativeai>=0.3.2
python-dotenv
